requirements.txt
# Core
pydantic>=2.10.4
pyyaml>=6.0.2
jsonlines>=4.0.0
fire>=0.7.0
datasketch>=1.6.5
torch>=2.5.1
accelerate>=1.2.1

# Hugging Face integrations
datasets>=3.2.0
transformers>=4.47.1
huggingface-hub>=0.27.0
safetensors>=0.4.5
zstandard>=0.23.0
tokenizers>=0.21.0

# Miscellaneous
tqdm>=4.67.1

# Quantization
bitsandbytes>=0.45.0
peft>=0.14.0

# Other
wandb>=0.19.1
tiktoken>=0.8.0
